pyproject.toml
[tool.poetry]
name = "petstore3"
version = "0.1.0"
description = ""
authors = ["Renat Devetyarov <devetyarov.renat@wb.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
requests = "^2.32.3"
allure-pytest = "^2.13.5"
faker = "^30.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
